+++ spotlight/src/main/java/com/takusemba/spotlight/SpotlightView.kt
@@ -1,6 +1,7 @@
 package com.takusemba.spotlight
 
 import android.animation.Animator
+import android.animation.AnimatorListenerAdapter
 import android.animation.ObjectAnimator
 import android.animation.TimeInterpolator
 import android.animation.ValueAnimator
@@ -10,14 +11,14 @@ import android.animation.ValueAnimator.ofFloat
 import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Paint
+import android.graphics.PointF
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffXfermode
 import android.util.AttributeSet
 import android.view.View
 import android.view.ViewGroup.LayoutParams.MATCH_PARENT
 import android.widget.FrameLayout
-import androidx.annotation.ColorRes
-import androidx.core.content.ContextCompat
+import androidx.annotation.ColorInt
 
 /**
  * [SpotlightView] starts/finishes [Spotlight], and starts/finishes a current [Target].
@@ -26,11 +27,11 @@ internal class SpotlightView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0,
-    @ColorRes backgroundColor: Int = R.color.background
+    @ColorInt backgroundColor: Int,
 ) : FrameLayout(context, attrs, defStyleAttr) {
 
   private val backgroundPaint by lazy {
-    Paint().apply { color = ContextCompat.getColor(context, backgroundColor) }
+    Paint().apply { color = backgroundColor }
   }
 
   private val shapePaint by lazy {
@@ -56,7 +57,7 @@ internal class SpotlightView @JvmOverloads constructor(
     val currentTarget = target
     val currentShapeAnimator = shapeAnimator
     val currentEffectAnimator = effectAnimator
-    if (currentTarget != null && currentEffectAnimator != null) {
+    if (currentTarget != null && currentEffectAnimator != null && currentShapeAnimator != null && !currentShapeAnimator.isRunning) {
       currentTarget.effect.draw(
           canvas = canvas,
           point = currentTarget.anchor,
@@ -109,23 +110,56 @@ internal class SpotlightView @JvmOverloads constructor(
   /**
    * Starts the provided [Target].
    */
-  fun startTarget(target: Target, listener: Animator.AnimatorListener) {
+  fun startTarget(target: Target) {
     removeAllViews()
     addView(target.overlay, MATCH_PARENT, MATCH_PARENT)
-    this.target = target
+    this.target = target.apply {
+      // adjust anchor in case where custom container is set.
+      val location = IntArray(2)
+      getLocationInWindow(location)
+      val offset = PointF(location[0].toFloat(), location[1].toFloat())
+      anchor.offset(-offset.x, -offset.y)
+    }
+    this.shapeAnimator?.removeAllListeners()
+    this.shapeAnimator?.removeAllUpdateListeners()
+    this.shapeAnimator?.cancel()
     this.shapeAnimator = ofFloat(0f, 1f).apply {
       duration = target.shape.duration
       interpolator = target.shape.interpolator
       addUpdateListener(invalidator)
-      addListener(listener)
+      addListener(object : AnimatorListenerAdapter() {
+        override fun onAnimationEnd(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+
+        override fun onAnimationCancel(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+      })
     }
+    this.effectAnimator?.removeAllListeners()
+    this.effectAnimator?.removeAllUpdateListeners()
+    this.effectAnimator?.cancel()
     this.effectAnimator = ofFloat(0f, 1f).apply {
+      startDelay = target.shape.duration
       duration = target.effect.duration
       interpolator = target.effect.interpolator
       repeatMode = target.effect.repeatMode
       repeatCount = INFINITE
       addUpdateListener(invalidator)
-      addListener(listener)
+      addListener(object : AnimatorListenerAdapter() {
+        override fun onAnimationEnd(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+
+        override fun onAnimationCancel(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+      })
     }
     shapeAnimator?.start()
     effectAnimator?.start()
@@ -136,15 +170,43 @@ internal class SpotlightView @JvmOverloads constructor(
    */
   fun finishTarget(listener: Animator.AnimatorListener) {
     val currentTarget = target ?: return
-    val currentShapeAnimator = shapeAnimator ?: return
-    shapeAnimator = ofFloat(currentShapeAnimator.animatedValue as Float, 0f).apply {
+    val currentAnimatedValue = shapeAnimator?.animatedValue ?: return
+    shapeAnimator?.removeAllListeners()
+    shapeAnimator?.removeAllUpdateListeners()
+    shapeAnimator?.cancel()
+    shapeAnimator = ofFloat(currentAnimatedValue as Float, 0f).apply {
       duration = currentTarget.shape.duration
       interpolator = currentTarget.shape.interpolator
       addUpdateListener(invalidator)
       addListener(listener)
+      addListener(object : AnimatorListenerAdapter() {
+        override fun onAnimationEnd(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+
+        override fun onAnimationCancel(animation: Animator) {
+          removeAllListeners()
+          removeAllUpdateListeners()
+        }
+      })
     }
+    effectAnimator?.removeAllListeners()
+    effectAnimator?.removeAllUpdateListeners()
     effectAnimator?.cancel()
     effectAnimator = null
     shapeAnimator?.start()
+  }
+
+  fun cleanup() {
+    effectAnimator?.removeAllListeners()
+    effectAnimator?.removeAllUpdateListeners()
+    effectAnimator?.cancel()
+    effectAnimator = null
+    shapeAnimator?.removeAllListeners()
+    shapeAnimator?.removeAllUpdateListeners()
+    shapeAnimator?.cancel()
+    shapeAnimator = null
+    removeAllViews()
   }
 }